refactor: Store 리팩토링 (#353)
