+++ src/main/java/project/trendpick_pro/domain/store/repository/StoreRepository.java
@@ -2,9 +2,7 @@ package project.trendpick_pro.domain.store.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.trendpick_pro.domain.store.entity.Store;
-
 import java.util.Optional;
-
 public interface StoreRepository extends JpaRepository<Store, Long> {
     Optional<Store> findByBrand(String brand);
 }
