+++ src/main/java/project/trendpick_pro/global/job/JobTestController.java
@@ -1,24 +1,19 @@
 package project.trendpick_pro.global.job;
 
-import jdk.jfr.Category;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.JobParametersInvalidException;
-import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
 import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
 import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.stereotype.Controller;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import project.trendpick_pro.global.job.JobConfig;
 
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("/job")
-public class JobController {
+public class JobTestController {
 
     private final JobConfig jobConfig;
 
@@ -28,9 +23,9 @@ public class JobController {
         try {
             jobConfig.performMakeRecommendProductJob();
         } catch (Exception e) {
-            throw new RuntimeException("Job 실행 실패");
+            throw new RuntimeException("추천 잡실행 실패");
         }
-        return "잡실행 성공";
+        return "추천 잡실행 성공";
     }
 
     @GetMapping("/order")
@@ -39,14 +34,14 @@ public class JobController {
         try {
             jobConfig.performCancelOrderJob();
         } catch (JobInstanceAlreadyCompleteException e) {
-            throw new RuntimeException(e);
+            throw new RuntimeException("주문삭제 잡실패");
         } catch (JobExecutionAlreadyRunningException e) {
-            throw new RuntimeException(e);
+            throw new RuntimeException("주문삭제 잡실패");
         } catch (JobParametersInvalidException e) {
-            throw new RuntimeException(e);
+            throw new RuntimeException("주문삭제 잡실패");
         } catch (JobRestartException e) {
-            throw new RuntimeException(e);
+            throw new RuntimeException("주문삭제 잡실패");
         }
-        return "성공";
+        return "주문삭제 잡성공";
     }
 }
